+++ Source/TowerDefense/Towers/InLineTower.h
@@ -86,7 +86,7 @@ protected:
 public:
 	FTowerObjectData* GetDataStruct();
 	void SetPlaced();
-	float TowerCost(float CurrentStarbucks);
+	int32 TowerCost(int32 CurrentStarbucks);
 
 	void SetSelected();
 	void DisableAttackRadiusDecal();
